1. Merge into once detected column type mismatch for primary
   key/partition key/precombined key we will error out.
2. Since MIT only requires column type matches for the 3 types of
   columns, for others spark-hudi did implicit type casting. Tests are
written to capture the exhaustive behavior of such handling.
3. For all the allowed implicit type casting, validate the behavior is
   consistent in case of insert and MIT.

Implementation of data type check

Followed today how we find the mapping of columns between source and
target table and apply to the partiton/primary/precombine column.

Refactoring other tests

1. For all cases where we change column type of target table creation
   statement, it is because the extra qualification of MIT requires data
type to match. Most of the test declare ts column as long yet MIT merges
int value to it. For simplicit ts column uses int now.
2. There is 1 test got removed where it validats we allow MIT with
   mismatching primary key data type in some cases. For example, we
allow long type in target and int type in source. This is no longer the
case with our new code change.

Allowing implicit casting of pk can leads to data correctness issue.
Here is an example:

        // Create target table with double primary key
        spark.sql(
          s"""
             |create table $targetTable (
             |  id double,
             |  name string,
             |  value_double double,
             |  ts long
             |) using hudi
             |location '${tmp.getCanonicalPath}/$targetTable'
             |tblproperties (
             |  type = '$tableType',
             |  primaryKey = 'id',
             |  preCombineField = 'ts'
             |)
         """.stripMargin)

        // Create source table with int primary key
        spark.sql(
          s"""
             |create table $sourceTable (
             |  id int,
             |  name string,
             |  value_double double,
             |  ts long
             |) using hudi
             |location '${tmp.getCanonicalPath}/$sourceTable'
             |tblproperties (
             |  type = '$tableType',
             |  primaryKey = 'id',
             |  preCombineField = 'ts'
             |)
         """.stripMargin)

        // Insert initial data into target table
        spark.sql(
          s"""
             |insert into $targetTable
             |select
             |  cast(1 as double) as id,
             |  'initial1' as name,
             |  1.1 as value_double,
             |  1000 as ts
         """.stripMargin)

        // Insert data into source table with int ids
        spark.sql(
          s"""
             |insert into $sourceTable
             |select
             |  1 as id,
             |  'updated1' as name,
             |  1.11 as value_double,
             |  1001 as ts
         """.stripMargin)

        // Perform merge operation
        spark.sql(
          s"""
             |merge into $targetTable t
             |using $sourceTable s
             |on t.id = s.id
             |when matched then update set *
             |when not matched then insert *
         """.stripMargin)

In the end we expect target table with 1 record
(1, updated1, 1.1, 1001)

but actually it is
(1, initial, 1.1, 1000)
(1, updated1, 1.1, 1001)

If target table id column is int, we got the expected behavior.

handle delete MIT action + exhaustive coverage

This commit addresses delete operation validation. It turns out that for
delete operation we don't need any specific handling, because:
- Unlike insert/update which contains assignment from source to target,
  delete operation does not assign values but just comparing them in the
  ON clause, we just need to ensure the comparison part accounts for
  type mismatches.
  Especially, the precombined key column is out of scope as it only takes
  effect when assignment happens. Similarily, if partition key/primary
  key are not involved in ON clause, we don't need to do any check.
- For condition clause, the recordKeyAttributeToConditionExpression
  variable already takes care of data type handling for both columns.
  Today's behavior is it will do best effort type casting to match
  source column data type to target, if cast succeeds everything works
  as expected, otherwise incompatible data type error is thrown.

Since exhaustive test cases are added validating MIT delete operations
in various type mismatch cases and no data correctness issue is spotted,
it is safe to relax the limitation of column type matching for delete
operations to include "convertible column types" cases. On contrast, MIT
insert and update enforce the data type to be exactly the same for all
3 types of columns given the various issues mentioned in the previous
commits.

Also added test coverage for MIT insert and update.

+++ hudi-spark-datasource/hudi-spark/src/test/resources/sql-statements.sql
@@ -63,10 +63,10 @@ select id, name, price, cast(dt as string) from h0_p;
 # CREATE TABLE
 
 create table h1 (
-  id bigint,
+  id int,
   name string,
   price double,
-  ts bigint
+  ts int
 ) using hudi
 options (
   type = '${tableType}',
@@ -79,10 +79,10 @@ location '${tmpDir}/h1';
 +----------+
 
 create table h1_p (
-  id bigint,
+  id int,
   name string,
   price double,
-  ts bigint,
+  ts int,
   dt string
 ) using hudi
 partitioned by (dt)
